Split some of the code out of package_info for generating a header.
There is one header we generate based on the modules found in the
package.

I was generating the file in package_info, but the file didn't appear in
conanmanifest and conan-upload complained.
This makes sense.
So I had to split and duplicate some code out of package_info and into
package.
It probably could have be done in a really nice elegant way, but I didn't.