qt/5.x.x: Fix opengl/system dependency to not be required when using es2
The opengl/system dependency is very misleading.
For Linux, it is a requirement for the legacy gl interface used for X11.
This is not necessarily required when using the es2 OpenGL backend and likely Wayland.

This PR only requires opengl/system on Linux when with_x11 is enabled.

Fixes #22116.

Cherry-picked from https://github.com/conan-io/conan-center-index/pull/22132