qt/6.x: drop cross_compile option
This option is somewhat broken and should be replaced with a global config option instead.
Libraries using Qt also need build tools and should use the same path as in that option, but the value of this option cannot be queried from `self.dependencies["qt"]"` in `build_requirements()`.

This option never had any effect on CCI anyway.

While a config option could be added instead, it would need extra handling in all consuming recipes, and it's not clear what benefit if any it would have over a `self.tool_requires("qt/...")`.