openfst: also permit building on Macos (#23592)
* openfst: also permit building on Macos

* openfst: remove explicit linkage against internal object files

On macOS the files referenced have the wrong type (MH_BUNDLE vs
MH_OBJECT or MH_DYLIB) and so the linker fails.  On Linux, the
additional files referenced cause dynamic link errors when trying to
execute the test_package when building with -o shared=True and -o
enable_const_fsts=True because the dynamic linker cannot find them.

Neither failure mode seems to persist if we simply use only the full
named library artifacts, so in the absence of a comment explaining the
logic or any reference in the git history, I'm inclined to simply remove
the additional .so references.

* Disable cross compiling and cleanup recipe

* Removed conan1 check and standarize conandata

---------

Co-authored-by: Rubén Rincón Blanco <git@rinconblanco.es>
Co-authored-by: PerseoGI <perseog@jfrog.com>