mimalloc: package minject.exe on Windows (#28359)
On Windows to ensure mimalloc works correctly it needs to be at the top of the import table in the final executable, mimalloc provides the minject.exe tool which allows to patch the executable. This packages it with the dlls.

Co-authored-by: Ernesto de Gracia Herranz <ernestodegraciah@gmail.com>
(cherry picked from commit baf342224a36ff78310603d3daf3ace9090de50a)