openmp: require that llvm-openmp version matches the compiler's internal LLVM version
To be on the safe side. To quote @jcar87:
"Conan Center users have already reported mismatches because new version of AppleClang will just fail to link or run with an anchored version of llvm-openmp"
https://github.com/conan-io/conan-center-index/pull/28440#discussion_r2390303992

Dropping OpenMP altogether on AppleClang is still unduly restrictive and does not match its active use in practice, though.